refactor: 코드 개선
- JSON 코드 분리
- force unwrapping 제거
- 기타 네이밍 변경
