+++ ContactManager/ContactManager/Models/ContactModel.swift
@@ -1,6 +1,6 @@
 import Foundation
 
-struct ContactModel: Codable {
+struct ContactDTO: Codable {
     let name: String
     let age: String
     let phoneNumber: String
