+++ ContactManager/ContactManager/SceneDelegate.swift
@@ -1,10 +1,3 @@
-//
-//  SceneDelegate.swift
-//  ContactManager
-//
-//  Created by imseonghyeon on 2023/10/04.
-//
-
 import UIKit
 
 class SceneDelegate: UIResponder, UIWindowSceneDelegate {
